update: modelの名前空間を変更
