+++ src/main/java/com/ushibutatory/zundoko/model/zundoko/Audience.java
@@ -1,4 +1,4 @@
-package com.ushibutatory.zundoko.model;
+package com.ushibutatory.zundoko.model.zundoko;
 
 /**
  * 観客クラス
