+++ src/main/java/com/ushibutatory/zundoko/model/zundoko/House.java
@@ -1,4 +1,4 @@
-package com.ushibutatory.zundoko.model;
+package com.ushibutatory.zundoko.model.zundoko;
 
 import java.util.ArrayList;
 import java.util.List;
