+++ src/main/java/com/ushibutatory/zundoko/model/zundoko/songs/SoranBushi.java
@@ -1,31 +1,31 @@
-package com.ushibutatory.zundoko.model.songs;
+package com.ushibutatory.zundoko.model.zundoko.songs;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 /**
- * ultra soul / B'z
+ * ソーラン節
  */
-public class Ultrasoul extends AbstractSong {
+public class SoranBushi extends AbstractSong {
 
     @Override
     public String getName() {
-        return "ultra soul";
+        return "ソーラン節";
     }
 
     @Override
     protected List<String> _createUsingPhrases() {
-        return new ArrayList<>(Arrays.asList("ウ", "ル", "ト", "ラ", "ソ"));
+        return new ArrayList<>(Arrays.asList("ヤーレン", "ソーラン"));
     }
 
     @Override
     protected List<Integer> _createCompletePhraseIndexes() {
-        return new ArrayList<>(Arrays.asList(0, 1, 2, 3, 4, 0, 1));
+        return new ArrayList<>(Arrays.asList(0, 1, 1, 0, 1, 1));
     }
 
     @Override
     protected String _createShoutPhrase() {
-        return "＼ハァイ！／";
+        return "＼ハイ！　ハイ！／";
     }
 }
