+++ src/main/java/com/ushibutatory/zundoko/model/zundoko/songs/Ultrasoul.java
@@ -1,31 +1,31 @@
-package com.ushibutatory.zundoko.model.songs;
+package com.ushibutatory.zundoko.model.zundoko.songs;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 /**
- * ずんどこ節
+ * ultra soul / B'z
  */
-public class ZundokoBushi extends AbstractSong {
+public class Ultrasoul extends AbstractSong {
 
     @Override
     public String getName() {
-        return "ズンドコ節";
+        return "ultra soul";
     }
 
     @Override
     protected List<String> _createUsingPhrases() {
-        return new ArrayList<>(Arrays.asList("ズン", "ドコ"));
+        return new ArrayList<>(Arrays.asList("ウ", "ル", "ト", "ラ", "ソ"));
     }
 
     @Override
     protected List<Integer> _createCompletePhraseIndexes() {
-        return new ArrayList<>(Arrays.asList(0, 0, 0, 0, 1));
+        return new ArrayList<>(Arrays.asList(0, 1, 2, 3, 4, 0, 1));
     }
 
     @Override
     protected String _createShoutPhrase() {
-        return "＼キ・ヨ・シ！／";
+        return "＼ハァイ！／";
     }
 }
