+++ src/main/java/com/ushibutatory/zundoko/model/zundoko/songs/ZundokoBushi.java
@@ -1,31 +1,31 @@
-package com.ushibutatory.zundoko.model.songs;
+package com.ushibutatory.zundoko.model.zundoko.songs;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 /**
- * ソーラン節
+ * ずんどこ節
  */
-public class SoranBushi extends AbstractSong {
+public class ZundokoBushi extends AbstractSong {
 
     @Override
     public String getName() {
-        return "ソーラン節";
+        return "ズンドコ節";
     }
 
     @Override
     protected List<String> _createUsingPhrases() {
-        return new ArrayList<>(Arrays.asList("ヤーレン", "ソーラン"));
+        return new ArrayList<>(Arrays.asList("ズン", "ドコ"));
     }
 
     @Override
     protected List<Integer> _createCompletePhraseIndexes() {
-        return new ArrayList<>(Arrays.asList(0, 1, 1, 0, 1, 1));
+        return new ArrayList<>(Arrays.asList(0, 0, 0, 0, 1));
     }
 
     @Override
     protected String _createShoutPhrase() {
-        return "＼ハイ！　ハイ！／";
+        return "＼キ・ヨ・シ！／";
     }
 }
