+++ src/main/java/com/ushibutatory/zundoko/service/BaseService.java
@@ -1,0 +1,13 @@
+package com.ushibutatory.zundoko.service;
+
+import org.springframework.stereotype.Service;
+
+/**
+ * 基底サービスクラス
+ */
+@Service
+public abstract class BaseService {
+
+    // Do nothing.
+
+}
